+++ api/driver/interface.go
@@ -1,4 +1,4 @@
-package interfaces
+package driver
 
 import "qp/internal/pkgdata"
