+++ internal/origins/drivers/deb/constants.go
