+++ internal/origins/drivers/pacman/constants.go
